renaming files and reorganizing directories within blogModule. Changed add button and removed unused imports
