+++ web/src/redcross/blogModule/BlogModule.ts
@@ -1,5 +1,5 @@
 import {NgModule} from "@angular/core";
-import {BlogComponent} from "./components/BlogComponent";
+import {BlogComponent} from "./components/BlogContainerComponent";
 import {RedcrossCommonModule} from "../commonModule/RedcrossCommonModule";
 import {BlogRoutingModule} from "./BlogRoutingModule";
 import {BlogPostComponent} from "./components/BlogPostComponent";
