+++ web/src/redcross/blogModule/BlogRoutingModule.ts
@@ -1,6 +1,6 @@
 import {NgModule} from "@angular/core";
 import {RouterModule, Routes} from "@angular/router";
-import {BlogComponent} from "./components/BlogComponent";
+import {BlogComponent} from "./components/BlogContainerComponent";
 import {BlogPostComponent} from "./components/BlogPostComponent";
 import {BlogDialogComponent} from "./components/BlogDialogComponent";
