+++ web/src/redcross/blogModule/components/Blog.less
@@ -10,5 +10,8 @@ div {
   position: fixed;
   right: 3vw;
   bottom: 5vh;
+  width: 8vw;
+  height: 4vh;
   background-color: lightgreen;
+  box-shadow: 10px 10px 5px #888888;
 }
