+++ web/src/redcross/blogModule/components/BlogPostComponent.ts
@@ -1,5 +1,5 @@
 import {Component, Input} from "@angular/core";
-import {Post} from "./Post";
+import {Post} from "../models/Post";
 
 @Component({
     selector: "blogPost",
