+++ web/src/redcross/blogModule/services/BlogDataService.ts
@@ -1,10 +1,10 @@
 import {Injectable} from "@angular/core";
 
-import {Post} from "./Post";
-import {Author} from "./Author";
+import {Post} from "../models/Post";
+import {Author} from "../models/Author";
 
 @Injectable()
-export class DiscussionDataService {
+export class BlogDataService {
     public getTestPosts(): Post[] {
                 return [testPost, testPost];
             }
